3 fixes:
-Change superbowl scores from ints to doubles to fix a site crash (why I'm up so late)
-Make weekly scores show one decimal, even if it's a 0. Also same with superbowl scores.
-If the user switches to a year with superbowl done, then back to the current year (which is during the superbowl but isn't finished yet), last year's winner/draft order shows up. I added a condition to check for this and hide it.

+++ HFL/GetJSON.aspx.cs
@@ -214,7 +214,7 @@ namespace HFL
                             break;
                         }
                     attrib = docNode.Attributes[HFLTeams[iTeam].sOwner];
-                    HFLTeams[iTeam].i15Points = Convert.ToInt32(attrib.Value);
+                    HFLTeams[iTeam].i15Points = Convert.ToDouble(attrib.Value);
                 }
             }
             if (xmlNL.Count == 16)
@@ -228,7 +228,7 @@ namespace HFL
                             break;
                         }
                     attrib = docNode.Attributes[HFLTeams[iTeam].sOwner];
-                    HFLTeams[iTeam].i16Points = Convert.ToInt32(attrib.Value);
+                    HFLTeams[iTeam].i16Points = Convert.ToDouble(attrib.Value);
                 }
             }
